Fix: do not load invalid hints for programs running on a segment != 0
Problem: the Cairo runner loads hints based on the PC offset. This means
that it cannot make a difference between the main program on segment 0
and any sub-program running on a segment != 0, ending up loading the main
program hints randomly when executing the sub-program.

Solution: disable hints for programs running on segments != 0. This
solution is similar to what is currently done for the instruction cache.